- Moved configuration classes into correct folder and adjusted namespaces accordingly
